+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## UNRELEASED
 
 ### Changes
 * Updates area UX to not display Add Content controls when a widget is focused
@@ -12,6 +12,10 @@
 ### Adds
 * Adds keyboard accessibility to Insert menu
 * Adds regex pattern feature for string fields.
+* Adds a warning if database queries involving relationships
+are made before the last `apostrophe:modulesRegistered` handler has fired.
+If you need to call Apostrophe's `find()` methods at startup,
+it is best to wait for the `@apostrophecms/doc:beforeReplicate` event.
 
 ## 3.48.0 (2023-05-26)